refactor(arch): move ArchitectureTest to platform-web, fix all 9 rules
- Move ArchitectureTest from platform-core to platform-web (sees all
  library module classes on classpath)
- Move ArchUnit dependency from platform-core to platform-web pom.xml
- Fix core dependency rule: target actual packages (model, service,
  analytics, export, search) instead of non-existent '..core..'
- Exclude OAuthRepository from repository rule (deferred: lives in
  security module until SecurityService decomposition)
- Exclude SecurityIntegrationTest from security rule (test class
  legitimately extends WebTest)
- Update AGENTS.md with stale classpath prevention guidance
