+++ platform-core/pom.xml
@@ -86,11 +86,6 @@
             <scope>test</scope>
         </dependency>
         <dependency>
-            <groupId>com.tngtech.archunit</groupId>
-            <artifactId>archunit-junit5</artifactId>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
             <groupId>io.insert-koin</groupId>
             <artifactId>koin-test-junit5</artifactId>
             <version>${koin.version}</version>
